requirements.txt
streamlit==1.32.2
pandas==2.2.1
plotly==5.20.0
openpyxl==3.1.2
xlsxwriter==3.1.9
fpdf2==2.8.2
